Revert commit 1cecdbab
Particle should not change the datatype of position and velocity, this
is the responsability of the client code. TrajectoryHDF5 now provides
float64 arrays for position and velocity by default (although it
should not cast integers).

+++ atooms/system/particle.py
@@ -23,11 +23,11 @@ class Particle(object):
         if position is None:
             self.position = numpy.zeros(_ndim)
         else:
-            self.position = numpy.asarray(position, dtype='float64')
+            self.position = numpy.asarray(position)
         if velocity is None:
             self.velocity = numpy.zeros(_ndim)
         else:
-            self.velocity = numpy.asarray(velocity, dtype='float64')
+            self.velocity = numpy.asarray(velocity)
 
     @property
     def diameter(self):
